+++ packages/terminal/src/browser/terminal-frontend-contribution.ts
@@ -33,10 +33,9 @@ import {
 } from '@theia/core/lib/browser';
 import { TabBarToolbarContribution, TabBarToolbarRegistry } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
 import { WidgetManager } from '@theia/core/lib/browser';
-import { TERMINAL_WIDGET_FACTORY_ID, TerminalWidgetFactoryOptions } from '@theia/xterm.js/lib/browser/terminal-widget-impl';
 import { TerminalKeybindingContexts } from './terminal-keybinding-contexts';
 import { TerminalService } from './terminal-service';
-import { TerminalWidgetOptions, TerminalWidget } from './terminal-widget';
+import { TerminalWidgetOptions, TerminalWidget, TERMINAL_WIDGET_FACTORY_ID, TerminalWidgetFactoryOptions } from './terminal-widget';
 // import { UriAwareCommandHandler } from '@theia/core/lib/common/uri-command-handler';
 import { FileSystem } from '@theia/filesystem/lib/common';
 import URI from '@theia/core/lib/common/uri';
@@ -138,7 +137,7 @@ export class TerminalFrontendContribution implements FrontendApplicationContribu
                 this.onDidCreateTerminalEmitter.fire(widget);
 
                 const clientOpsToRestore = this.termClientsToRestore.get(widget.id);
-                if (clientOpsToRestore) {
+                if (clientOpsToRestore) { // todo use time instead of domId.
                     const client = this.newTerminalClient(clientOpsToRestore, widget as TerminalWidget);
                     if (clientOpsToRestore.terminalId) {
                         client.attach(clientOpsToRestore.terminalId);
