+++ packages/xterm-js/src/browser/terminal-frontend-module.ts
@@ -14,11 +14,10 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import { ContainerModule, Container } from 'inversify';
-import { TerminalWidgetImpl, TERMINAL_WIDGET_FACTORY_ID } from './terminal-widget-impl';
-import { TerminalWidget, TerminalWidgetOptions } from '@theia/terminal/lib/browser/terminal-widget';
+import { ContainerModule } from 'inversify';
+import { TerminalWidgetImpl } from './terminal-widget-impl';
+import { TerminalWidget } from '@theia/terminal/lib/browser/terminal-widget';
 import { bindTerminalPreferences } from './terminal-preferences';
-import { WidgetFactory } from '@theia/core/lib/browser';
 
 import '../../src/browser/terminal.css';
 import 'xterm/lib/xterm.css';
@@ -28,25 +27,4 @@ export default new ContainerModule(bind => {
     bindTerminalPreferences(bind);
 
     bind(TerminalWidget).to(TerminalWidgetImpl).inTransientScope();
-
-    let terminalNum = 0;
-    bind(WidgetFactory).toDynamicValue(ctx => ({
-        id: TERMINAL_WIDGET_FACTORY_ID,
-        createWidget: (options: TerminalWidgetOptions) => {
-            const child = new Container({ defaultScope: 'Singleton' });
-            child.parent = ctx.container;
-            const counter = terminalNum++;
-            const domId = options.id || 'terminal-' + counter;
-            const widgetOptions: TerminalWidgetOptions = {
-                title: 'Terminal ' + counter,
-                useServerTitle: true,
-                destroyTermOnClose: true,
-                ...options
-            };
-            child.bind(TerminalWidgetOptions).toConstantValue(widgetOptions);
-            child.bind('terminal-dom-id').toConstantValue(domId);
-
-            return child.get(TerminalWidget);
-        }
-    }));
 });
